create DeadlineTask and refactor EventTask
