+++ DoLah/Task.cpp
@@ -1,7 +1,6 @@
 #include "Task.h"
 
 namespace DoLah {
-    using namespace std::chrono;
 
     Task::Task() {}
     Task::~Task() {}
@@ -37,22 +36,5 @@ namespace DoLah {
     int Task::getId() {
         return this->id;
     }
-    EventTask::EventTask() { }
-    EventTask::~EventTask() { }
 
-    void EventTask::setEndDate(system_clock::time_point datetime) {
-        this->endDate = datetime;
-    }
-
-    system_clock::time_point EventTask::getEndDate() {
-        return this->endDate;
-    }
-
-    void EventTask::setStartDate(system_clock::time_point datetime) {
-        this->startDate = datetime;
-    }
-
-    system_clock::time_point EventTask::getStartDate() {
-        return this->startDate;
-    }
 }
